+++ apps/openmw/mwmechanics/spellpriority.cpp
@@ -16,6 +16,7 @@
 
 #include "creaturestats.hpp"
 #include "spellcasting.hpp"
+#include "spellresistance.hpp"
 #include "weapontype.hpp"
 #include "combat.hpp"
 #include "summoning.hpp"
